recommender/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.5.0

# Supabase & config
supabase>=2.10.0
python-dotenv>=1.0.0

# Data
pandas>=2.1.0
numpy>=2.0.0
scipy>=1.14.0

# ML / Models
scikit-learn>=1.5.0
implicit>=0.7.0
torch>=2.1.0
sentence-transformers>=2.2.0
transformers>=4.35.0
open-clip-torch>=2.23.0

# Indexing
faiss-cpu>=1.7.4

# Image handling
Pillow>=10.0.0
requests>=2.31.0

# Dashboard
streamlit>=1.28.0

# Utilities
beautifulsoup4>=4.12.0
joblib>=1.3.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Containerisation
docker>=6.1.0
